+++ resources/views/admin/subjects/form.blade.php
@@ -46,21 +46,23 @@
             {!! Form::submit($content, ['class' => 'btn btn-primary']) !!}
             {!! Form::close() !!}
             <a href="{{ route('subjects.index') }}" class="btn btn-info">{{ __('Back') }}</a>
-            @if (isset($subject))
-                @if (!in_array($subject->id, $subjectsHasScores->toArray()))
-                    {!! Form::open([
-                        'method' => 'DELETE',
-                        'route' => ['subjects.destroy', $subject->id],
-                        'style' => 'display:inline;',
-                    ]) !!}
-                    {!! Form::button(__('Delete'), [
-                        'type' => 'submit',
-                        'class' => 'btn btn-danger',
-                        'onclick' => 'return confirm("' . __('Are you sure?') . '")',
-                    ]) !!}
-                    {!! Form::close() !!}
+            @can('destroy_subject')
+                @if (isset($subject))
+                    @if (!in_array($subject->id, $subjectsHasScores->toArray()))
+                        {!! Form::open([
+                            'method' => 'DELETE',
+                            'route' => ['subjects.destroy', $subject->id],
+                            'style' => 'display:inline;',
+                        ]) !!}
+                        {!! Form::button(__('Delete'), [
+                            'type' => 'submit',
+                            'class' => 'btn btn-danger',
+                            'onclick' => 'return confirm("' . __('Are you sure?') . '")',
+                        ]) !!}
+                        {!! Form::close() !!}
+                    @endif
                 @endif
-            @endif
+            @endcan
         </div>
     </div>
 @endsection
